requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.3
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.3
jupyterlab-server==1.0.6
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
oauthlib==3.1.0
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
praw==6.4.0
prawcore==1.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
Pygments==2.4.2
pyrsistent==0.15.6
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
pywin32==227
pywinpty==0.5.5
pyzmq==18.1.1
requests==2.22.0
requests-oauthlib==1.3.0
Send2Trash==1.5.0
six==1.13.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
tweepy==3.8.0
update-checker==0.16
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
zipp==0.6.0
